OKay so the tabs are already created just need to fix the. delivery charge system, and implement in the checkout page.
The admin panel has the system mentioned.

If both the seller user and customer user has the same Thana/Upazilla in their addresses then the charge will be Intra Thana/Upazilla determined charge by the admin in admin panel, and for Intra Distrcit but not in the same Thana/Upazilla the charge will be Intra District charge and If nothing is the same, then It will be charged as Inter District charge determined by the admin in admin panel. Same division or not doesn't matter.

Implement this system in the checkout page while customer user checkout of the product.