Show the delivery charge in the cart page as well.
If the customer user is not logged in or didn't submit an address then in cart page It will show "Add Address to calculate" or something.

and for now on to upload a project, a user must add their addresses. If the seller user has not added his address then his upload product page will redirect to profile settings and in profile settings, there will be a red notice that " Complete your profile to sell items" and "add an address to sell item", "add a withdrwal method to sell item" in their boxes.