In the mobile nav bar, category icon.
Clicking in this icon will show up a half screen covering popup from down hiding navbar. This will consist categories, parent categories. 2 categories in a row. If there are a lot of categories then the user may scroll but the pop up not exced half of the screen, there will also be a cross button to hide that, additionally clicling anywhere in the csreen beside the category pop up will hide It, It will go down.

So first page of that down to halfpage up popup will be category and If someone selects the category It will show It's sub categories alongside a subcategory called "All" which is not available as sub category but It will show the page of the parent category having all the proucts no matter the sub category. But selecting any sub category will only show the product of that particular sub category