So there is another thing. In the order alongside the delivery status, there will be another status which is payment status. It can be not paid or paid. User's my earing will be increase or add the money valued the product value only after the order status is delivered and paid. not before that.
and admin panel will have another option which is comission tab.

In comission tab admin will fix how much comission he wants from each of the categories. For that category admin dashboard will show income just on the parcentage of the total sell as comission.

For example user has set a prodcut in books category for 100 taka, admin from admin panel decided to charge 30% comission in all the books category products. So, after delivery is complete and status is paid, the seller user my earning panel withdrwable income will be 70 BDT and admin. dashboard's admin's income will be 30BDT. Although in admin dashboard there will be other matrics such as total sells, showing the total sells of the platforms, total delivery charge collected, etc.

the commission will be only charged on product value not on delivery charge. And the seller user is not suppoed to get the delivery charge or any parcentage.