Okay so category will have images, and these images can be updated from the admin panel, sub categories will have images well and these can be edited or updated or added from admin panel sub category tabs.
I need something similiar to the brand tab, which will not be brand instead, for each category that will be different name, that admin can choose, and they will be only category related.

For example, I have created, the category called Books, sub category, admission preparation books, and then I want to make a brand like "Pubisher", and "Author", naming that as well as adding the data for example, for books category, there will be publishers and Authors,named Prothom Alo and Mr. Rahim.

For different categories, the name "Brand" will be set by admin and will be updated.

This kind of sections can be added as much as admin wants for example admin wants to make 3 brands like category for Category books, "Author", "Publication", "Language" and he can populate them.

When admin or user add a product for sell and select category, the sub category drop down box will only show sub categories related to that category and there will be new boxes of that brand like names will appear, and user will select that, which will be related to the category and upload these products.

When selecting a category in the website, or a product page, user can filter them by categories or brand like sectiosn or categories