The comission logic is not working properly.
Again I revise myself.

The comission will only be applied on the product sub total not the delivery charge.

The comission by category adjusted by the admin will be effected after the order is delivered and the payment status is paid.

for example admin has decided to adjust the comission for books as 30% from the admin panel.

Someone buys a book from the category and paid 100 BDT. admin will get 30 in his dashboard, and user who is seller will get 70 in their waithdrawable earning or the Pending Earnings.

Now when I tried this with 100 taka product, with a 30% comission in the books category. Unfortunately user is getting only 40 BDT and admin is getting something else. not what I expected, user 70 and admin 30.

Fix the comission issue.