There will be only one Dhaka City, no North or South.
As the system delivery charge is based on the Intra Upazilla, Intra District (but not Intra Upazilla) and Inter District basis. Dhaka City north and Dhaka city south may occure Intra District charge instead of Intra Upazilla charge, which is not currect, Dhaka City North and South includes in the Intra Upazilla Delivery charge in our system.

and you have worked with a lot of thinsg related to the location, Is the system of delivery still Intact?