Make the following changes:
,

Find this element:

Relevant files:

    - https://6000-firebase-studio-1748277312119.cluster-iktsryn7xnhpexlu6255bftka4.cloudworkstations.dev/_next/static/chunks/node_modules_cc6f26b6._.js
    - https://6000-firebase-studio-1748277312119.cluster-iktsryn7xnhpexlu6255bftka4.cloudworkstations.dev/_next/static/chunks/src_1e1711e5._.js
    - https://6000-firebase-studio-1748277312119.cluster-iktsryn7xnhpexlu6255bftka4.cloudworkstations.dev/_next/static/chunks/src_1e1711e5._.js

And change it as follows:

In the mobile version this header becomes too big and missmatch the header section and makes It messy. In this place only a small square size logo is suited. So please in this place do not use the logo inseat use favicon of the website. Admin should be able to upload or update favicon from admin panel which admin can right now I guess, good. Just replace this place with favicon, which doesn't require much space